For ENG-4098, robustify the barrier used for snapshot setup
Instead of initializing with a variable number of things and getting it wrong, always initialize to the number of sites on the host.

Have all site threads even those involved in rejoin arrive at the barrier when they are needed by checking for waiters (don't judge me!)

This gives us the isolation from changes in rejoin state that we need.