For ENG-4098, counter for when a snapshot is done is broken.
Fix it by recording exactly which sites are involved using a synchronized set of the snapshot site processor objects.

This requires an exit barrier for snapshot setup because it is possible for a site to finish setting and finish the snapshot before another site has even registered itself in the set. This causes the terminator to run early.

Also add a missing barrier check to RejoinProducer live rejoin, while it is spinning it needs to monitor the barrier.