For ENG-4141, pass UniqueId into EE for export, and pass SpHandle to EE instead of TxnId.
Previously the transaction id was passed in which doesn't work in IV2 with export because the multi-part transaction id is passed in and that can be lower then the single part id causing export to think the transaction id went backwards. Instead pass export the SpHandle for the site doing the insert along with the unique id for the transaction and use the unique id in the export data. This yields user visible output that is most similar to what existed pre-IV2.

Also refactor all references to transaction id to be references to an SpHandle to reflect the change. That ends up touching a bunch of files and is the bulk of the commit.

By fixing the input to export TestExportOpsEventsTestConnectionToRejoiningCluster now passes. It failed because it did a multi-part insert into an export table causing export
to complain about the lower multi-part txnid.