ENG-5393: Fix upper bound iterator in elastic index.
The range is now all inclusive on both ends, the upper bound iterator needs to
include all tuples with the given hash. Since the elastic index key is a
composite key with hash and pointer address, upper bound has to pass the max
pointer value to get the proper iterator.

Also changed the pointer used in the composite key to unsigned integer, because
comparing less than of two pointers in C++ is undefined.