ENG-5601: Fix findBlock in persistent table and the COW context.
When the COW context looks for the block that contains a tuple, it's checking if
the block start address + block size is smaller than the tuple address, if not,
the tuple is thought to be in the block. If two blocks are adjacent and the
tuple is the first one in the latter block, it would actually think that the
first block contains the tuple. The check should be smaller than or equal to,
because the start address is inclusive.

persistenttable.findBlock() also has the same problem. I consolidated the three
blocks in CopyOnWriteContext.cpp and persistenttable.h into one.