Add a flag to SnapshotCompletionInterest to indicate the fate of the snapshot.
Here's what I think the contract is. The completion interest will fire on all
finished snapshots, regardless of whether or not the snapshot succeeded, as long
as all nodes have done the work. If node failures happened during the snapshot,
it won't fire.

To determine if the snapshot actually succeeded, check the didSucceed flag in
the event.