For ENG-6247, update export format to be more space efficient so it can be shared with DR
The main change is to use smaller sizes for smaller integer types instead of always using a long.

For decimals instead of using a string prefix with scale and precision in bytes. Right now this is always 12 and 16, but in the future it serialization could trim to the smallest necessary scale and precision and other scales and precisions can be supported.

The prefixes are large enough to exceed the supported representions of postgres and mysql, and postgres's is quite large.