ENG-6769: Remove stats tracking from JDBC client connection.
There is already a stats tracker inside the Java client. The JDBC client
connection stats object is not reset after warmup in the VoltKV benchmark.

Instead of resetting it, I changed it to use the stats context provided by the
Java client just like all the other benchmarks. This will make the stats summary
at the end more accurate.