ENG-7849: Rollback BEGIN TXN entry in the binary log.
It was possible that when a transaction rolls back, it doesn't roll back
the BEGIN TXN entry, leaving it orphaned in the binary log. This causes
problems like inconsistent unique IDs for the same sequence number
downstream.

Returning the correct start USO for the undo actions to correctly
rollback all entries generated by a transaction.