Fix nonsensical parameter passing to TableFactory::getPersistentTable() in c++ unit tests.
Apparently, we've been passing a pointer to a DR stream object to a
boolean parameter in TableFactory::getPersistentTable(). That parameter
has nothing to do with DR. It compiled on g++ and earlier versions of
clang because it does a pointer to boolean auto conversion. The newer
version of clang started complaining about this.

This has made all subsequent parameters to be misaligned. Fixed the
problem by removing the DR stream parameter and all subsequent parameter
values if they match the defaults.