Clear existing children when connecting subqueries to plan
When I was writing some planner tests, I was encountering failed
assertions because a scan plan nodes had more than one child.  I
traced the cause to the method `connectChildrenBestPlans` which seems
to have been added as part of in/exists/scalar subquery support.

I wasn't able to recreate this issue using sqlcmd (I tried both ad hoc
and procedures), but seems like this issue might manifest in a
real-world scenario?

The added test was failing (in master) before the change to
`PlanAssembler`.