Calculate non-determinism of aggregates of floating point types.
If some aggregate functions are applied to floating point parameters the
result may be non-deterministic, due to floating point numeric
differences.  We need to causes errors or warnings in this case.  This
is complicated by the fact that the error messages need to propagated up
from the aggregate call expression through the containing expressions,
into the subquery statement object.  From there it may need to propagate
into a join tree or in ParsedUnionStatement objects if the expression is
in a subquery of a joined table or in a subquery of a union or
intersection expression.