ENG-9383: Fix durability notification after rejoin.
On rejoin, the durability listener won't be notified until the first SP
and MP transactions are logged. This means that if the cluster is idle,
nothing is made durable. However, we do know that anything on disk prior
to the node failing must have been made durable by other nodes.

- Notify the durability listeners even if the last txn was a sysproc
  that doesn't require durability. Previously, we were just letting it
  through without notifying any listener.

- Propogate the last unique ID from the master SPI to the rejoining SPI
  using the Iv2LogFaultMessage, which is the very first message a
  rejoining SPI will get. Then initialize the durability listener with
  this unique ID.

Change-Id: Iadfa4c86d53314ee06e4c912cecbbbd20cdd4e80