Replace Boost regular expressions with PCRE2 regular expressions.
We replaced dependence on the Boost regular expression
library with dependence on the PCRE2 library.  The latter is much
smaller and easier to build than Boost.  It is alleged that Boost regex
can be built without building all of Boost via bjam and b2, but we have
not investigated this.

The PCRE2 sources are added as one single tarball, to reduce our github
repository size.

The regular expression matching function, regexp_position, was reworked
to use the PCRE2 library.  Some C++ unit tests were added as well.

https://issues.voltdb.com/browse/ENG-9556