Clean up lookups for USING and Ambiguous columns.
In ENG-9367 we change the lookup rules to reject queries which had
references to ambiguous columns.  This caused a number of queries in
plannerTester to fail, as they were ambiguous.

However, our implementation in ENG-9367 was flawed, in that we rejected
references like R1.C in "select R1.C from R1 join R2 using(C)".  HSQLDB
called these unreferenceable, and so we did too.  This is apparently
incorrect.  I checked with mysql and postgres.  Both consider "select C
from R1 join R2 using(C), R3" to be ambiguous when R1, R2 and R3 all
have columns named C.  They are happy with the same query and schemata
with "C" in the select list replaced by "R1.C" or "R2.C".  We have added
tests for all these cases.

The relevant section of the 2003 SQL standard seems to be section 7.7
paragraph (7).