ENG-9919, use uniqueId to match sentinel and first MP fragment on replay.
In ENG-9779, we don't reuse txnId of MP fragments from CL on replay to avoid
interleaving txnId for MP fragments and SP txn. But ReplaySequencer checks to
see if txnId matches between sentinel and MP fragments before it releases MP
fragments, so here we use uniqueId instead to match the sentinel and fragments.

Change-Id: I70fe249471a55413e7ebd6be827460d5788a3454