ENG-9977, bug fix, DR consumer doesn't send sync snapshot request when DR is disabled at startup and enabled later.
if the database starts without DR enabled, the DR consumer won't be notified during the leader election.
After DR is enabled, DR consumer doesn't have the knowledge of who the locally led partitions are. So
all the dispatchers think they are not the leaders and wait for other to send the sync snapshot request.

ENG-9977, decouple ConsumerDRGateway from initiator, remove the DummyConsumerDRGateway.