Execute an EE fragment in a unit test.
This adds an example of a unit test function which executes a fragment
of a plan. There was an example of this already, but it was not
completely fleshed out. We simply create an object of a subclass of
ExecutionEngine, another of a subclass of Topend, and join them
together. With some careful buffer management we can execute a fragment
and arrange for the output table to be accessible. We can then validate
the output for correctness.

ENG-10022