Parse dates better in sqlcmd.
The documented forms for VoltDB dates are 'yyyy-mm-dd HH:MM.sss',
'yyyy-MM-dd HH:MM' and 'yyyy-MM-dd'. The previous pattern did not
work for the latter two. This commit fixes this by trying all three
patterns. There seems to be no way to specify that a part of the
pattern is optional.

I changed the parsing strategy, and just use the one for TimestampType.
This
gives us microsecond granularity, which is what we seem to be
advertising,
and it parses all the formats we advertise.

In this commit I also added a fix to ENG-10369. It was really obvious,
and easier to fix it than to describe how to fix it in the ticket.

https://issues.voltdb.com/browse/ENG-10031