ENG-10486 fail partitioned cluster earlier (#3682)
When a cluster is partitioned due to network failure, we would like to detect that the cluster is not KSafe (all partitions are present) before performing any partition repair process. This patch addresses this issue.

IsKSafe is called on the zookeeper thread but makes a zookeeper call (which is not allowed. Move the check to the executor.

Clean up the test case that verifies fault logs after hosts go away.
Also fix possible NPE in CommandLogReplayState.
And a little bit of code formatting cleanup.