ENG-10617 (#3690)
When truncation snapshots are taken the absolute path is specified. For LocalCluster, this means that all hosts will write their snapshots to the same directory rather than the directory designated as the root for each
host. This works fine if each snapshot instance contains data for all partitions but if the cluster is multi-host and not k-safe, multiple snapshots are required to complete a RECOVER.

This patch alters VoltFile (which is already LocalCluster aware) to convert file and directory requests to a different host directory to the correct host directory.